Add multiple table file sink to base
