+++ seatunnel-connectors-v2/connector-file/connector-file-local/src/main/java/org/apache/seatunnel/connectors/seatunnel/file/local/sink/LocalFileSink.java
@@ -17,89 +17,16 @@
 
 package org.apache.seatunnel.connectors.seatunnel.file.local.sink;
 
-import org.apache.seatunnel.api.common.JobContext;
 import org.apache.seatunnel.api.configuration.ReadonlyConfig;
-import org.apache.seatunnel.api.serialization.DefaultSerializer;
-import org.apache.seatunnel.api.serialization.Serializer;
-import org.apache.seatunnel.api.sink.SeaTunnelSink;
-import org.apache.seatunnel.api.sink.SinkAggregatedCommitter;
-import org.apache.seatunnel.api.sink.SinkWriter;
-import org.apache.seatunnel.api.sink.SupportMultiTableSink;
 import org.apache.seatunnel.api.table.catalog.CatalogTable;
-import org.apache.seatunnel.api.table.type.SeaTunnelRow;
 import org.apache.seatunnel.connectors.seatunnel.file.config.FileSystemType;
-import org.apache.seatunnel.connectors.seatunnel.file.config.HadoopConf;
-import org.apache.seatunnel.connectors.seatunnel.file.hadoop.HadoopFileSystemProxy;
 import org.apache.seatunnel.connectors.seatunnel.file.local.config.LocalFileHadoopConf;
-import org.apache.seatunnel.connectors.seatunnel.file.local.sink.writter.LocalFileSinkWriter;
-import org.apache.seatunnel.connectors.seatunnel.file.sink.commit.FileAggregatedCommitInfo;
-import org.apache.seatunnel.connectors.seatunnel.file.sink.commit.FileCommitInfo;
-import org.apache.seatunnel.connectors.seatunnel.file.sink.commit.FileSinkAggregatedCommitter;
-import org.apache.seatunnel.connectors.seatunnel.file.sink.config.FileSinkConfig;
-import org.apache.seatunnel.connectors.seatunnel.file.sink.state.FileSinkState;
-import org.apache.seatunnel.connectors.seatunnel.file.sink.writer.WriteStrategy;
-import org.apache.seatunnel.connectors.seatunnel.file.sink.writer.WriteStrategyFactory;
+import org.apache.seatunnel.connectors.seatunnel.file.sink.BaseMultipleTableFileSink;
 
-import java.util.List;
-import java.util.Optional;
-
-public class LocalFileSink
-        implements SeaTunnelSink<
-                        SeaTunnelRow, FileSinkState, FileCommitInfo, FileAggregatedCommitInfo>,
-                SupportMultiTableSink {
-
-    private final HadoopConf hadoopConf;
-    private final HadoopFileSystemProxy hadoopFileSystemProxy;
-    private final FileSinkConfig fileSinkConfig;
-    private final WriteStrategy writeStrategy;
-    private String jobId;
+public class LocalFileSink extends BaseMultipleTableFileSink {
 
     public LocalFileSink(ReadonlyConfig readonlyConfig, CatalogTable catalogTable) {
-        this.hadoopConf = new LocalFileHadoopConf();
-        this.fileSinkConfig =
-                new FileSinkConfig(readonlyConfig.toConfig(), catalogTable.getSeaTunnelRowType());
-        this.writeStrategy =
-                WriteStrategyFactory.of(fileSinkConfig.getFileFormat(), fileSinkConfig);
-        this.hadoopFileSystemProxy = new HadoopFileSystemProxy(hadoopConf);
-        this.writeStrategy.setSeaTunnelRowTypeInfo(catalogTable.getSeaTunnelRowType());
-    }
-
-    @Override
-    public void setJobContext(JobContext jobContext) {
-        this.jobId = jobContext.getJobId();
-    }
-
-    @Override
-    public SinkWriter<SeaTunnelRow, FileCommitInfo, FileSinkState> restoreWriter(
-            SinkWriter.Context context, List<FileSinkState> states) {
-        return new LocalFileSinkWriter(writeStrategy, hadoopConf, context, jobId, states);
-    }
-
-    @Override
-    public Optional<SinkAggregatedCommitter<FileCommitInfo, FileAggregatedCommitInfo>>
-            createAggregatedCommitter() {
-        return Optional.of(new FileSinkAggregatedCommitter(hadoopConf));
-    }
-
-    @Override
-    public SinkWriter<SeaTunnelRow, FileCommitInfo, FileSinkState> createWriter(
-            SinkWriter.Context context) {
-        return new LocalFileSinkWriter(writeStrategy, hadoopConf, context, jobId);
-    }
-
-    @Override
-    public Optional<Serializer<FileCommitInfo>> getCommitInfoSerializer() {
-        return Optional.of(new DefaultSerializer<>());
-    }
-
-    @Override
-    public Optional<Serializer<FileAggregatedCommitInfo>> getAggregatedCommitInfoSerializer() {
-        return Optional.of(new DefaultSerializer<>());
-    }
-
-    @Override
-    public Optional<Serializer<FileSinkState>> getWriterStateSerializer() {
-        return Optional.of(new DefaultSerializer<>());
+        super(new LocalFileHadoopConf(), readonlyConfig, catalogTable);
     }
 
     @Override
